Renamed the ui folder as primitives
