+++ src/App/src/components/LoginForm.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
-import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
+import { Button } from '@/components/primitives/button';
+import { Input } from '@/components/primitives/input';
+import { Label } from '@/components/primitives/label';
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/primitives/card';
 import { User, Lock, Envelope } from '@phosphor-icons/react';
 
 interface LoginFormProps {
