+++ src/App/src/components/ProductRecommendation.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { ShoppingCart, Star } from '@phosphor-icons/react';
-import { Button } from '@/components/ui/button';
-import { Card, CardContent } from '@/components/ui/card';
-import { Badge } from '@/components/ui/badge';
+import { Button } from '@/components/primitives/button';
+import { Card, CardContent } from '@/components/primitives/card';
+import { Badge } from '@/components/primitives/badge';
 import { Product } from '@/lib/types';
 
 interface ProductRecommendationProps {
