+++ src/App/src/components/primitives/alert-dialog.tsx
@@ -4,7 +4,7 @@ import { ComponentProps } from "react"
 import * as AlertDialogPrimitive from "@radix-ui/react-alert-dialog"
 
 import { cn } from "@/lib/utils"
-import { buttonVariants } from "@/components/ui/button"
+import { buttonVariants } from "@/components/primitives/button"
 
 function AlertDialog({
   ...props
