+++ src/App/src/components/primitives/calendar.tsx
@@ -4,7 +4,7 @@ import ChevronRight from "lucide-react/dist/esm/icons/chevron-right"
 import { DayPicker } from "react-day-picker"
 
 import { cn } from "@/lib/utils"
-import { buttonVariants } from "@/components/ui/button"
+import { buttonVariants } from "@/components/primitives/button"
 
 function Calendar({
   className,
