+++ src/App/src/components/primitives/command.tsx
@@ -11,7 +11,7 @@ import {
   DialogDescription,
   DialogHeader,
   DialogTitle,
-} from "@/components/ui/dialog"
+} from "@/components/primitives/dialog"
 
 function Command({
   className,
