+++ src/App/src/components/primitives/form.tsx
@@ -12,7 +12,7 @@ import {
 } from "react-hook-form"
 
 import { cn } from "@/lib/utils"
-import { Label } from "@/components/ui/label"
+import { Label } from "@/components/primitives/label"
 
 const Form = FormProvider
