+++ src/App/src/components/primitives/slider.tsx
