+++ src/App/src/components/primitives/toggle-group.tsx
@@ -5,7 +5,7 @@ import * as ToggleGroupPrimitive from "@radix-ui/react-toggle-group"
 import { type VariantProps } from "class-variance-authority"
 
 import { cn } from "@/lib/utils"
-import { toggleVariants } from "@/components/ui/toggle"
+import { toggleVariants } from "@/components/primitives/toggle"
 
 const ToggleGroupContext = createContext<
   VariantProps<typeof toggleVariants>
